fix: CRUD Product

+++ RestAPI/src/main/java/com/restApi/RestAPI/controllers/ProeductsController.java
@@ -1,16 +1,22 @@
 package com.restApi.RestAPI.controllers;
 
 import com.restApi.RestAPI.dto.outputDTO.ResponseDTOOutput;
+import com.restApi.RestAPI.model.product.ImageStore;
 import com.restApi.RestAPI.model.product.Products;
 import com.restApi.RestAPI.services.ProductsService;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/products")
@@ -32,7 +38,15 @@ public class ProeductsController {
     }
 
     @DeleteMapping("/{productId}")
-    public ResponseEntity<ResponseDTOOutput> deleteUser(@PathVariable Long productId) {
+    public ResponseEntity<ResponseDTOOutput> deleteUser(@PathVariable Long productId, HttpServletRequest request) {
+        String userRole = (String) request.getAttribute("userRole");
+        if(!Objects.equals(userRole, "admin")){
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg("Unauthorized role");
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
+        }
+
         ResponseDTOOutput status = productsService.deleteById(productId);
         if ("success".equals(status.getStatus())) {
             return ResponseEntity.ok(status);
@@ -42,24 +56,75 @@ public class ProeductsController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<ResponseDTOOutput> createProductions(@RequestBody Products inputUser) {
-        ResponseDTOOutput status = productsService.createProduct(inputUser);
-        if ("success".equals(status.getStatus())) {
-            return ResponseEntity.ok(status);
-        }else {
-            return ResponseEntity.badRequest().body(status);
+    public ResponseEntity<ResponseDTOOutput> createProduct(
+            @RequestParam("title") String title,
+            @RequestParam("description") String description,
+            @RequestParam("price") Float price,
+            @RequestParam("duration") Integer duration,
+            @RequestParam("stock") Integer stock,
+            @RequestParam(value = "image", required = false) MultipartFile image,
+            @RequestParam(value = "imageStoreName", required = false) String imageName,
+            HttpServletRequest request)
+    {
+        String userRole = (String) request.getAttribute("userRole");
+        if(!Objects.equals(userRole, "admin")){
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg("Unauthorized role");
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
+        }
+
+        try {
+            // Panggil service untuk menyimpan produk
+            ResponseDTOOutput response = productsService.createProduct(title, description, price, duration, stock, image, imageName);
+            return ResponseEntity.ok(response);
+        } catch (IOException e) {
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg("Failed to save image");
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        } catch (Exception e) {
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg(e.getMessage());
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
         }
     }
 
     @PutMapping("/update/{productId}")
     public ResponseEntity<ResponseDTOOutput> updateProductions(
-            @RequestBody Products inputUser,
-            @PathVariable Long productId) {
-        ResponseDTOOutput status = productsService.updateProduct(inputUser, productId);
-        if ("success".equals(status.getStatus())) {
-            return ResponseEntity.ok(status);
-        } else {
-            return ResponseEntity.badRequest().body(status);
+            @PathVariable Long productId,
+            @RequestParam("title") String title,
+            @RequestParam("description") String description,
+            @RequestParam("price") Float price,
+            @RequestParam("duration") Integer duration,
+            @RequestParam("stock") Integer stock,
+            @RequestParam(value = "prevImage", required = false) String prevImage,
+            @RequestParam(value = "image", required = false) MultipartFile image,
+            @RequestParam(value = "imageStoreName", required = false) String imageName,
+            HttpServletRequest request)
+    {
+        String userRole = (String) request.getAttribute("userRole");
+        if(!Objects.equals(userRole, "admin")){
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg("Unauthorized role");
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
+        }
+
+        try {
+            ResponseDTOOutput response = productsService.updateProduct(title, description, price, duration, stock, prevImage, image, imageName, productId);
+            return ResponseEntity.ok(response);
+        } catch (IOException e) {
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg("Failed to save image");
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        } catch (Exception e) {
+            ResponseDTOOutput errorResponse = new ResponseDTOOutput();
+            errorResponse.setMsg(e.getMessage());
+            errorResponse.setStatus("failed");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
         }
     }
